First dodgy pass code version
This version does a first initial inversion, and then smooths the
output. It's a first pass to what's coming afterwards. At present,
the code has a large number of specific choices that need looking
into: it should be made more generic in terms of model parameters
and boundaries, but this gives an answer.