Reader for Sentinel 1 LMU netCDF files
The files from LMU come in the annyoing netCDF format. This class
basically allows one to read them in and subset them to a region
of interest. Seems to work, assuming the LMU format is stable, but
has a number of missing features, such as the chunking implementation.
This can probably be done in a similar way to the S2 data.