Operation on images
The current code includes `kaska_sar.py`, which uses the S2
inversion data to provide a (smoothed) estimate of LAI or
LAI*Cab to the inverter. Currently, the inversion works on a
per-pixel basis, and is super-inefficient. I have also observed
that the WCM is unable to fit quite a few pixels. There's a need
for bounds in the solver, as well as testing which solver is the
most efficient.