Set up problem interface
This commit finishes setting up the problem for S2 inversions.
Basically, the user only selects a set of options, calls a function
and the magic happens. The code is designed to split the problem
into spatial windows and invert them sequentially, but is also able
to exploit dask to run this in parallel.