Added logging infrastructure
I have now added logging both to a file and console. The code is
consistent for all submodules. It's spitting things twice to the
console, so there's some issue in adding handlers, but other than
that, it appears to be working OK.

There are lots of changes to all files, but basically most of them
are just simplifying things.